+++ src/Client/Ranges.hpp
@@ -8,11 +8,14 @@ namespace qolmod
     {
         std::vector<Range> ranges = {};
 
-        void sort();
         bool getEnable(float value, bool def, bool* inAnyRange = nullptr);
+
         void addRange(float min, float max, bool enable);
         void addRange(qolmod::Range range);
+
+        void sort();
         void clear();
+        bool isEmpty();
 
         matjson::Value save();
         void load(const matjson::Value& value);